2014-10-06 21:10 UTC+0200 Przemyslaw Czerpak (druzus/at/poczta.onet.pl)   * src/rtl/gtxwc/gtxwc.c     ! fixed multiple input context initialization in HB_GTI_FONTSEL.       Thanks to Rolf for the information.
  * contrib/hbnetio/netiocli.c
  * contrib/hbwin/wapi_winbase.c
  * src/rtl/iousr.c
  * src/vm/cmdarg.c
    * pacified MSVC warnings

+++ contrib/hbwin/wapi_winbase.c
@@ -299,7 +299,7 @@ HB_FUNC( WAPI_FORMATMESSAGE )
       }
    }
    else
-      dwFlags = ~FORMAT_MESSAGE_ALLOCATE_BUFFER;
+      dwFlags = ( DWORD ) ~FORMAT_MESSAGE_ALLOCATE_BUFFER;
 
    if( dwFlags & FORMAT_MESSAGE_ALLOCATE_BUFFER )
       lpBuffer = ( LPTSTR ) &lpAllocBuff;
